requirements.txt
fastapi
uvicorn
langchain
crewai
groq
python-dotenv
sqlalchemy
langchain-groq
python-multipart
requests
